Move business logic from widget to repository
